Add example brainrender html export
